requirements.txt
asttokens==2.4.1
attrs==23.2.0
branca==0.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
folium==0.16.0
fonttools==4.49.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nbformat==5.9.2
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tornado==6.4
traitlets==5.14.1
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xyzservices==2023.10.1
